+++ app/src/main/java/lol/terabrendon/houseshare2/data/remote/dto/ExpensePartDto.kt
@@ -4,5 +4,5 @@ data class ExpensePartDto(
     val id: Long,
     val expenseId: Long,
     val userId: Long,
-    val partAmount: Double,
+    val partAmount: Long,
 )
